+++ Protecs/src/main/java/uk/ac/ucl/protecs/sim/Params.java
@@ -12,8 +12,10 @@ import java.util.List;
 import java.util.Map;
 
 import uk.ac.ucl.protecs.objects.*;
-import uk.ac.ucl.protecs.objects.Location.LocationCategory;
-import uk.ac.ucl.protecs.objects.Person.OCCUPATION;
+import uk.ac.ucl.protecs.objects.hosts.Person;
+import uk.ac.ucl.protecs.objects.hosts.Person.OCCUPATION;
+import uk.ac.ucl.protecs.objects.locations.Location;
+import uk.ac.ucl.protecs.objects.locations.Location.LocationCategory;
 
 public class Params {
