+++ Protecs/src/test/java/uk/ac/ucl/protecs/objects/diseases/CovidSpuriousSymptomTesting.java
@@ -3,10 +3,10 @@ package uk.ac.ucl.protecs.objects.diseases;
 import org.junit.Assert;
 import uk.ac.ucl.protecs.helperFunctions.*;
 import uk.ac.ucl.protecs.helperFunctions.helperFunctions.NodeOption;
+import uk.ac.ucl.protecs.objects.hosts.Person;
 
 import org.junit.Test;
 
-import uk.ac.ucl.protecs.objects.Person;
 import uk.ac.ucl.protecs.behaviours.CoronavirusBehaviourFramework.CoronavirusBehaviourNodeTitle;
 import uk.ac.ucl.protecs.sim.WorldBankCovid19Sim;
 import uk.ac.ucl.protecs.sim.WorldBankCovid19Sim.DISEASE;
